dice: Remove deprecated solution
Per DICE Layering Architecture specification, it mentions
the key must depend on CDI.
The currnet solution uses the hash of CDI digest as the seed
to derive the key pair meets the requirement.
Therefore, the original solution that uses the cdi as the
seed of rand() and then use the result of rand() as the seed
to derive key pair can be removed.

Signed-off-by: Steven Lee <steven_lee@aspeedtech.com>
Change-Id: I04d5ff9d903fd9ecae396e5c76e211d89c8d5863