dice: adjust Basic Constraints and Key usage extensions
This code change is used to generate an Alias certificate for testing
with spdm-emu. The X509v3 extensions in the Alias certificate is like
below.

X509v3 extensions:
    X509v3 Key Usage:
        Digital Signature, Non Repudiation, Key Encipherment
    X509v3 Extended Key Usage:
        TLS Web Client Authentication
    X509v3 Authority Key Identifier:
        96:D6:32:16:7F:BB:E0:DA:BF:9D:EA:F5:47:EC:51:4E:24:CD:A9:29
    X509v3 Basic Constraints: critical
        CA:FALSE, pathlen:1

CA bit is set to false because Alias certificate is a leaf
certificate and it will not be used to sign a CSR.
Digital Signature, Non Repudiation and Key Encipherment are set
because Alias certificate will be used to sign a content during the
protocol handshaking.

Note : when user signs the DevID certificate, the CA bit should be
asserted because the DevID certificate will be used to sign the Alias
certificate.

Signed-off-by: Kane-Chen-AS <kane_chen@aspeedtech.com>
Change-Id: If7e828baf5c5131c6a1ca7d81eff0d6231b75e1c